+++ Sources/MapboxDirections/MapMatching/MapMatchingResponse.swift
@@ -1,4 +1,7 @@
 import Foundation
+#if canImport(FoundationNetworking)
+import FoundationNetworking
+#endif
 
 public struct MapMatchingResponse {
     public let httpResponse: HTTPURLResponse?
@@ -50,6 +53,5 @@ extension MapMatchingResponse: Codable {
         
         tracepoints = try container.decodeIfPresent([Tracepoint?].self, forKey: .tracepoints)
         matches = try container.decodeIfPresent([Match].self, forKey: .matches)
-
     }
 }