+++ Sources/MapboxDirections/Route.swift
@@ -1,4 +1,7 @@
+import Foundation
+#if canImport(CoreLocation)
 import CoreLocation
+#endif
 import Turf
 
 /**
@@ -15,9 +18,10 @@ open class Route: DirectionsResult {
      - parameter shape: The roads or paths taken as a contiguous polyline.
      - parameter distance: The route’s distance, measured in meters.
      - parameter expectedTravelTime: The route’s expected travel time, measured in seconds.
+     - parameter typicalTravelTime: The route’s typical travel time, measured in seconds.
      */
-    public override init(legs: [RouteLeg], shape: LineString?, distance: CLLocationDistance, expectedTravelTime: TimeInterval) {
-        super.init(legs: legs, shape: shape, distance: distance, expectedTravelTime: expectedTravelTime)
+    public override init(legs: [RouteLeg], shape: LineString?, distance: CLLocationDistance, expectedTravelTime: TimeInterval, typicalTravelTime: TimeInterval? = nil) {
+        super.init(legs: legs, shape: shape, distance: distance, expectedTravelTime: expectedTravelTime, typicalTravelTime: typicalTravelTime)
     }
     
     /**
@@ -36,6 +40,7 @@ extension Route: Equatable {
         return lhs.routeIdentifier == rhs.routeIdentifier &&
             lhs.distance == rhs.distance &&
             lhs.expectedTravelTime == rhs.expectedTravelTime &&
+            lhs.typicalTravelTime == rhs.typicalTravelTime &&
             lhs.speechLocale == rhs.speechLocale &&
             lhs.responseContainsSpeechLocale == rhs.responseContainsSpeechLocale &&
             lhs.legs == rhs.legs &&