+++ Tests/MapboxDirectionsTests/AnnotationTests.swift
@@ -1,11 +1,14 @@
 import XCTest
-#if !SWIFT_PACKAGE
+#if !os(Linux)
+#if SWIFT_PACKAGE
+import OHHTTPStubsSwift
+#endif
 import OHHTTPStubs
 @testable import MapboxDirections
 
 class AnnotationTests: XCTestCase {
     override func tearDown() {
-        OHHTTPStubs.removeAllStubs()
+        HTTPStubs.removeAllStubs()
         super.tearDown()
     }
     
@@ -23,9 +26,9 @@ class AnnotationTests: XCTestCase {
         ]
         
         stub(condition: isHost("api.mapbox.com")
-            && containsQueryParams(queryParams)) { _ in 
-                let path = Bundle(for: type(of: self)).path(forResource: "annotation", ofType: "json")
-                return OHHTTPStubsResponse(fileAtPath: path!, statusCode: 200, headers: ["Content-Type": "application/json"])
+                && containsQueryParams(queryParams)) { _ in
+            let path = Bundle.module.path(forResource: "annotation", ofType: "json")
+            return HTTPStubsResponse(fileAtPath: path!, statusCode: 200, headers: ["Content-Type": "application/json"])
         }
         
         let options = RouteOptions(coordinates: [