+++ Tests/MapboxDirectionsTests/MatchTests.swift
@@ -1,18 +1,26 @@
 import XCTest
-#if !SWIFT_PACKAGE
+#if canImport(CoreLocation)
+import CoreLocation
+#else
+import Turf
+#endif
+#if !os(Linux)
 import OHHTTPStubs
+#if SWIFT_PACKAGE
+import OHHTTPStubsSwift
+#endif
 #endif
 @testable import MapboxDirections
 
 class MatchTests: XCTestCase {
     override func tearDown() {
-        #if !SWIFT_PACKAGE
-        OHHTTPStubs.removeAllStubs()
+        #if !os(Linux)
+        HTTPStubs.removeAllStubs()
         #endif
         super.tearDown()
     }
     
-    #if !SWIFT_PACKAGE
+    #if !os(Linux)
     func testMatch() {
         let expectation = self.expectation(description: "calculating directions should return results")
         let locations = [CLLocationCoordinate2D(latitude: 32.712041, longitude: -117.172836),
@@ -26,8 +34,8 @@ class MatchTests: XCTestCase {
         stub(condition: isHost("api.mapbox.com")
             && isMethodGET()
             && pathStartsWith("/matching/v5/mapbox/driving")) { _ in
-                let path = Bundle(for: type(of: self)).path(forResource: "match", ofType: "json")
-                return OHHTTPStubsResponse(fileAtPath: path!, statusCode: 200, headers: ["Content-Type": "application/json"])
+                let path = Bundle.module.path(forResource: "match", ofType: "json")
+                return HTTPStubsResponse(fileAtPath: path!, statusCode: 200, headers: ["Content-Type": "application/json"])
         }
         
         var response: MapMatchingResponse!
@@ -115,8 +123,8 @@ class MatchTests: XCTestCase {
         stub(condition: isHost("api.mapbox.com")
             && isMethodGET()
             && pathStartsWith("/matching/v5/mapbox/driving")) { _ in
-                let path = Bundle(for: type(of: self)).path(forResource: "null-tracepoint", ofType: "json")
-                return OHHTTPStubsResponse(fileAtPath: path!, statusCode: 200, headers: ["Content-Type": "application/json"])
+                let path = Bundle.module.path(forResource: "null-tracepoint", ofType: "json")
+                return HTTPStubsResponse(fileAtPath: path!, statusCode: 200, headers: ["Content-Type": "application/json"])
         }
         
         var response: MapMatchingResponse!