+++ Tests/MapboxDirectionsTests/RoutableMatchTests.swift
@@ -1,11 +1,14 @@
 import XCTest
-#if !SWIFT_PACKAGE
+#if !os(Linux)
 import OHHTTPStubs
+#if SWIFT_PACKAGE
+import OHHTTPStubsSwift
+#endif
 @testable import MapboxDirections
 
 class RoutableMatchTest: XCTestCase {
     override func tearDown() {
-        OHHTTPStubs.removeAllStubs()
+        HTTPStubs.removeAllStubs()
         super.tearDown()
     }
     
@@ -22,13 +25,14 @@ class RoutableMatchTest: XCTestCase {
         stub(condition: isHost("api.mapbox.com")
             && isMethodGET()
             && pathStartsWith("/matching/v5/mapbox/driving")) { _ in
-                let path = Bundle(for: type(of: self)).path(forResource: "match", ofType: "json")
-                return OHHTTPStubsResponse(fileAtPath: path!, statusCode: 200, headers: ["Content-Type": "application/json"])
+                let path = Bundle.module.path(forResource: "match-polyline6", ofType: "json")
+                return HTTPStubsResponse(fileAtPath: path!, statusCode: 200, headers: ["Content-Type": "application/json"])
         }
         
         var routeResponse: RouteResponse!
         
         let matchOptions = MatchOptions(coordinates: locations)
+        matchOptions.shapeFormat = .polyline6
         matchOptions.includesSteps = true
         matchOptions.routeShapeResolution = .full
         for waypoint in matchOptions.waypoints[1..<(locations.count - 1)] {
@@ -56,7 +60,7 @@ class RoutableMatchTest: XCTestCase {
         let route = routeResponse.routes!.first!
         XCTAssertNotNil(route)
         XCTAssertNotNil(route.shape)
-        XCTAssertEqual(route.shape!.coordinates.count, 18)
+        XCTAssertEqual(route.shape!.coordinates.count, 19)
         XCTAssertEqual(route.routeIdentifier, nil)
         
         let waypoints = routeResponse.waypoints!
@@ -93,7 +97,7 @@ class RoutableMatchTest: XCTestCase {
         XCTAssertEqual(step.finalHeading, 340)
         
         XCTAssertNotNil(step.shape)
-        XCTAssertEqual(step.shape!.coordinates.count, 4)
+        XCTAssertEqual(step.shape!.coordinates.count, 5)
         let coordinate = step.shape!.coordinates.first!
         XCTAssertEqual(round(coordinate.latitude), 33)
         XCTAssertEqual(round(coordinate.longitude), -117)