+++ Tests/MapboxDirectionsTests/RouteStepTests.swift
@@ -1,5 +1,9 @@
 import XCTest
+#if canImport(CoreLocation)
 import CoreLocation
+#else
+import Turf
+#endif
 @testable import MapboxDirections
 
 class RoadTests: XCTestCase {
@@ -113,6 +117,7 @@ class RouteStepTests: XCTestCase {
             "ref": "CA 24",
             "weight": 2.5,
             "duration": 2.5,
+            "duration_typical": 2.369,
             "name": "Grove Shafter Freeway (CA 24)",
             "pronunciation": "ˈaɪˌfoʊ̯n ˈtɛn",
             "distance": 24.5,
@@ -139,6 +144,7 @@ class RouteStepTests: XCTestCase {
             XCTAssertEqual(step.instructions, "Keep right onto CA 24")
             XCTAssertEqual(step.codes, ["CA 24"])
             XCTAssertEqual(step.expectedTravelTime, 2.5)
+            XCTAssertEqual(step.typicalTravelTime, 2.369)
             XCTAssertEqual(step.names, ["Grove Shafter Freeway"])
             XCTAssertEqual(step.phoneticNames, ["ˈaɪˌfoʊ̯n ˈtɛn"])
             XCTAssertEqual(step.distance, 24.5)
@@ -177,6 +183,7 @@ class RouteStepTests: XCTestCase {
                 "instruction": "Turn left onto Adalbertstraße",
             ],
             "duration": 59.1,
+            "duration_typical": 45.0,
             "distance": 236.9,
             "driving_side": "right",
             "weight": 59.1,
@@ -231,6 +238,9 @@ class RouteStepTests: XCTestCase {
             XCTAssertEqual(step.shape?.coordinates.last?.latitude ?? 0, 52.506794, accuracy: 1e-5)
             XCTAssertEqual(step.shape?.coordinates.last?.longitude ?? 0, 13.42326, accuracy: 1e-5)
             
+            XCTAssertEqual(step.expectedTravelTime, 59.1)
+            XCTAssertEqual(step.typicalTravelTime, 45.0)
+            
             XCTAssertNoThrow(encodedStepData = try encoder.encode(step))
             XCTAssertNotNil(encodedStepData)
             
@@ -267,6 +277,63 @@ class RouteStepTests: XCTestCase {
             XCTAssertNotNil(encodedStepJSON)
             
             XCTAssertEqual(encodedStepJSON?["pronunciation"] as? String, "ˈaɪˌfoʊ̯n ˈtɛn; ˈaɪˌfoʊ̯n ˈtɛnz")
+        }
+    }
+    
+    func testRouteStepTypicalTravelTime() {
+        let typicalTravelTime = 2.5
+        
+        let route = RouteStep(transportType: .automobile,
+                              maneuverLocation: CLLocationCoordinate2D(latitude: 0, longitude: 0),
+                              maneuverType: .turn,
+                              instructions: "",
+                              drivingSide: .left,
+                              distance: 2.0,
+                              expectedTravelTime: 2.0,
+                              typicalTravelTime: typicalTravelTime)
+
+        XCTAssertEqual(route.typicalTravelTime, typicalTravelTime)
+    }
+    
+    func testIncidentsCoding() {
+        let path = Bundle.module.path(forResource: "incidents", ofType: "json")
+        let filePath = URL(fileURLWithPath: path!)
+        let data = try! Data(contentsOf: filePath)
+        let options = RouteOptions(coordinates: [
+            CLLocationCoordinate2D(latitude: 37.78, longitude: -122.42),
+            CLLocationCoordinate2D(latitude: 38.91, longitude: -77.03),
+        ])
+        
+        let decoder = JSONDecoder()
+        decoder.userInfo[.options] = options
+        decoder.userInfo[.credentials] = DirectionsCredentials(accessToken: "foo", host: URL(string: "http://sample.website"))
+        let result = try! decoder.decode(RouteResponse.self, from: data)
+        
+        let routes = result.routes
+        let route = routes!.first!
+
+        
+        // Encode and decode the route securely.
+        
+        let encoder = JSONEncoder()
+        encoder.userInfo[.options] = options
+        encoder.outputFormatting = [.prettyPrinted]
+        
+        var jsonData: Data?
+        XCTAssertNoThrow(jsonData = try encoder.encode(route))
+        XCTAssertNotNil(jsonData)
+    
+        if let jsonData = jsonData {
+            var newRoute: Route?
+            XCTAssertNoThrow(newRoute = try decoder.decode(Route.self, from: jsonData))
+            XCTAssertNotNil(newRoute)
+            
+            XCTAssert(newRoute!.legs.first!.incidents!.first!.kind == Incident.Kind.miscellaneous)
+            XCTAssert(newRoute!.legs.first!.incidents![0].lanesBlocked!.contains(.right))
+            XCTAssertNil(newRoute!.legs.first!.incidents![1].lanesBlocked)
+            XCTAssert(newRoute!.legs.first!.incidents![2].lanesBlocked!.isEmpty)
+            XCTAssert(newRoute!.legs.first!.incidents![2].shapeIndexRange == 810..<900)
+            XCTAssert(newRoute!.legs.first!.incidents!.first! == route.legs.first!.incidents!.first!)
         }
     }
 }