+++ Tests/MapboxDirectionsTests/RouteTests.swift
@@ -1,4 +1,10 @@
 import XCTest
+import Foundation
+#if canImport(CoreLocation)
+import CoreLocation
+#else
+import Turf
+#endif
 @testable import MapboxDirections
 
 class RouteTests: XCTestCase {
@@ -10,6 +16,7 @@ class RouteTests: XCTestCase {
                     "summary": "West 6th Avenue Freeway, South University Boulevard",
                     "weight": 1346.3,
                     "duration": 1083.4,
+                    "duration_typical": 1483.262,
                     "steps": [],
                     "distance": 17036.8,
                 ],
@@ -17,6 +24,7 @@ class RouteTests: XCTestCase {
             "weight_name": "routability",
             "weight": 1346.3,
             "duration": 1083.4,
+            "duration_typical": 1483.262,
             "distance": 17036.8,
         ]
         let routeData = try! JSONSerialization.data(withJSONObject: routeJSON, options: [])
@@ -33,10 +41,10 @@ class RouteTests: XCTestCase {
         decoder.userInfo[.options] = options
         XCTAssertNoThrow(route = try decoder.decode(Route.self, from: routeData))
         
-        let expectedLeg = RouteLeg(steps: [], name: "West 6th Avenue Freeway, South University Boulevard", distance: 17036.8, expectedTravelTime: 1083.4, profileIdentifier: .automobileAvoidingTraffic)
+        let expectedLeg = RouteLeg(steps: [], name: "West 6th Avenue Freeway, South University Boulevard", distance: 17036.8, expectedTravelTime: 1083.4, typicalTravelTime: 1483.262, profileIdentifier: .automobileAvoidingTraffic)
         expectedLeg.source = options.waypoints[0]
         expectedLeg.destination = options.waypoints[1]
-        let expectedRoute = Route(legs: [expectedLeg], shape: nil, distance: 17036.8, expectedTravelTime: 1083.4)
+        let expectedRoute = Route(legs: [expectedLeg], shape: nil, distance: 17036.8, expectedTravelTime: 1083.4, typicalTravelTime: 1483.262)
         XCTAssertEqual(route, expectedRoute)
         
         if let route = route {
@@ -138,7 +146,13 @@ class RouteTests: XCTestCase {
         }
 
         // Then - voiceLocale key is present and its value is nil
-
-        XCTAssertTrue(json.contains { (key, value) in key == "voiceLocale" && value == nil })
+        // swift-corelibs-foundation’s implementation of JSONSerialization uses NSNull to represent a JSON null value.
+        XCTAssertTrue(json.contains { (key, value) in key == "voiceLocale" && (value == nil || value is NSNull) })
+    }
+    
+    func testRouteTypicalTravelTime() {
+        let typicalTravelTime = 2.5
+        let route = Route(legs: [], shape: nil, distance: 2.0, expectedTravelTime: 3.0, typicalTravelTime: typicalTravelTime)
+        XCTAssertEqual(route.typicalTravelTime, typicalTravelTime)
     }
 }